Now Messages Saves on DB from Frontend
post request add in the sendMsg function to save messages. and to set userstatus array now there is a function to reduce code. and 'offline' socket emit event is removed from return, because I used return for when user refresh page but refreshing page does not means that user logout, and 'offline' event was login out user from messaging service.